pyproject.toml
[tool.poetry]
name = "fun-telegram-bot"
version = "0.1.0"
description = "Telegram bot just for sake of fun"
authors = ["Andriy Sydorenko <sidorenkoandrij217@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "3.6.0"
alembic = "1.13.1"
SQLAlchemy = "2.0.30"
pytz = "2024.1"
pillow = "^10.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
package-mode = false

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.git
    | \.env
    | \.ini
    | alembic
  )/
  | \.pyi?$
)
'''
